requirements.txt
langchain
ctransformers
sentence-transformers==2.2.2
pinecone-client==2.2.4
flask
numpy
huggingface_hub
pandas
matplotlib
seaborn
pypdf
python-dotenv
-e .
torch
